Fixed a couple of small bugs as per Akaya.
Specifically, you can now add Frames directly to the HUD, and they will
draw.

And Chat Compounds with the tabs on the Left and Right will now properly
size their MiniConsoles; they were reversed in the logic (a Left-tabbed
Chat was using a MiniConsole meant for a Right-tabbed Chat and vice
versa).