update Vyzor package.path mangling to put the Vyzor paths first.
This was recommended by SlySven from the Mudlet forums: "In the Vyzor loader, visible as the Vyzor/Vyzor script in the Mudlet editor window, I note that you temporarily change the Lua package.path to include a couple of extra directories to check to load the Vyzor subsystem from. However you add the directories to the END of the package.path not the beginning. Now I've made some changes to the System Lua loader code (so it can find {possibly even read-only} shared copies of the Mudlet and Geyser lua code shipped with the Mudlet source code - as would happen with a Linux System Distribution or Administrator install to the /usr/share or /usr/local/share directory trees respectively.) As a side effect something has happened to the package.path that your loader is receiving and I think I might be munged something wrongly - however if I put the extra paths that Vyzor inserts at the BEGINNING then all is well"