expand OrderedTable functionality
The purpose of this modification was to embed count information in the table itself, instead of using a second variable to manage what is, practically speaking, data internal to the table.
As part of this, I turned some of the remaining key-value tables into OrderedTables, regardless of any need for ordering. This may, in fact, be somewhat of a hack. Also, moved iteration out
of the metatable and into a number of separate functions on the returned proxy in order to provide better call semantics.