Create config.lua
This fixes an issue where Mudlet has to use the zip name for the package name - thus the latest released "Vyzor-3.3.3.zip" gets imported as "vyzor-333". This subsequently breaks vyzor loader, which expects the folder name to be called ``vyzor`` in ``pcall(require, "vyzor")`` so it can find ``init.lua`` within.